requirements.txt

Flask==2.1.1
flask-cors==3.0.10
transformers==4.18.0
torch==1.11.0
Pillow==9.1.0
gunicorn==20.1.0
